subDomainsBrute/requirements.txt
dnspython==1.16.0
gevent==1.4.0
greenlet==0.4.15
